set current branch from env or git rev-parse
Use TRAVIS_BRANCH if set.

There is no reason to assume that the subprocess command would fail
during a travis build, but since the branch has already been determined
it might be slightly safter/faster to use that.

In development, subprocess is fine.

The git command will return "HEAD" if a specific commit or tag is checked out
unfortunately. But then the links in the documentation will just be broken.