Make shell commands easier to copy
Example shell commands were using the console formatting which looks nicer, especially when you have the command and then all the output following the command. However, it requires placing a $ in front of each command, which makes it difficult to copy and paste.

For all the commands that we are displaying output, we use console with the $. For the rest of the commands we swich to shell so its easier to copy.